Proof of concept
I hard-coded in the similarity matrix - still doing some work on the SQL
database and auto-populating any updated ratings (it's based on a cosine
similarity technique).

This demo also includes the upfront quiz for people to rate varieties of
wine, which fills in for wines that have not been rated (with the
exception of the wine being viewed, of course).  Everything updates in
real time (with alerts to confirm that things work).